+++ Common/src/main/java/fuzs/pickupnotifier/client/gui/entry/ItemDisplayEntry.java
@@ -5,10 +5,11 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import fuzs.pickupnotifier.PickUpNotifier;
 import fuzs.pickupnotifier.config.ClientConfig;
 import net.minecraft.network.chat.Component;
+import net.minecraft.world.ContainerHelper;
+import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.ItemStack;
 
 public class ItemDisplayEntry extends DisplayEntry {
-
     private final ItemStack stack;
 
     public ItemDisplayEntry(ItemStack stack, int amount) {
@@ -20,18 +21,24 @@ public class ItemDisplayEntry extends DisplayEntry {
     @Override
     protected Component getEntryName() {
 
-        return PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries ? this.getStackName() : this.stack.getHoverName();
-    }
-
-    private Component getStackName() {
-
-        return this.stack.getItem().getName(this.stack);
+        return PickUpNotifier.CONFIG.get(ClientConfig.class).behavior.combineEntries ? this.stack.getItem().getName(this.stack) : this.stack.getHoverName();
     }
 
     @Override
     public boolean mayMergeWith(DisplayEntry other) {
 
-        return other instanceof ItemDisplayEntry && this.stack.getItem() == ((ItemDisplayEntry) other).stack.getItem() && this.getStackName().equals(((ItemDisplayEntry) other).getStackName());
+        return other instanceof ItemDisplayEntry itemDisplayEntry && this.sameItem(itemDisplayEntry.stack);
+    }
+
+    private boolean sameItem(ItemStack other) {
+
+        return this.stack.getItem() == other.getItem() && this.stack.getItem().getName(this.stack).equals(other.getItem().getName(other));
+    }
+
+    @Override
+    protected int getInventoryCount(Inventory inventory) {
+
+        return ContainerHelper.clearOrCountMatchingItems(inventory, this::sameItem, Integer.MAX_VALUE, true);
     }
 
     @Override
@@ -42,6 +49,12 @@ public class ItemDisplayEntry extends DisplayEntry {
         modelViewStack.scale(scale, scale, 1.0F);
         RenderSystem.applyModelViewMatrix();
         this.minecraft.getItemRenderer().renderAndDecorateItem(this.stack, posX, posY);
+
+        if (PickUpNotifier.CONFIG.get(ClientConfig.class).display.displayAmount.sprite()) {
+
+            DisplayEntryRenderHelper.renderGuiItemDecorations(this.minecraft.font, this.getDisplayAmount(), posX, posY);
+        }
+
         modelViewStack.popPose();
         RenderSystem.applyModelViewMatrix();
     }